添加视频类新闻JSON转模型

+++ QQNews/QQNews/News(新闻)/Model/视频/ETTCardModel.swift
@@ -16,6 +16,6 @@ class ETTCardModel: NSObject
     var icon:String?
     var uin:String?
     var intro:String?
-    var msgEntry:NSInteger?
+    var msgEntry:Int?
     
 }
