+++ QQNews/QQNews/News(新闻)/Model/视频/ETTVideoChannelModel.swift
@@ -12,7 +12,7 @@ class ETTVideoChannelModel: NSObject
 {
     var egid:String?
     var eid:String?
-    var openSupport:NSInteger?
+    var openSupport:Int?
     var video:ETTVideoModel?
     
 }
